Spring: Quickfix: Multiplied the length by zero in angular spring
The scene was multiplying the length of the angular spring by zero when the angle
was updated. This is fixed by passing 1 on the angle. I've also added a few lines to
detect if zero is passed to the length, at which point it will add to the log (if in
debug mode) and set the length adjustment to 1 since I may have made that mistake
elsewhere in the code.